+++ components/layout/navbar/Navbar.tsx
@@ -1,11 +1,27 @@
 'use client';
 
+import { SegmentNode } from '@/types/page';
 import NavbarDetail from './NavbarDetail';
 import NavbarRoot from './NavbarRoot';
-import useNavbar from './useNavbar';
+import { useState } from 'react';
+
+export type State =
+  | {
+      // 접힌 상태
+      type: 'closed';
+    }
+  | {
+      // 펼쳐진 상태
+      type: 'expanded';
+    }
+  | {
+      // 세부 페이지 보이는 상태
+      type: 'hovered';
+      segmentNode: SegmentNode;
+    };
 
 export default function Navbar() {
-  const [state, setState] = useNavbar();
+  const [state, setState] = useState({ type: 'closed' });
 
   const handleMouseLeave = () => {
     // 세부 페이지가 보이고 있다면 닫기
